AS7-2912 Fix deserialization classloader logic, now that Infinispan provides basic support via AdvancedCache.with(ClassLoader)
