+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/DefaultEmbeddedCacheManager.java
@@ -22,18 +22,35 @@
 
 package org.jboss.as.clustering.infinispan;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.security.AccessController;
+import java.security.PrivilegedAction;
+import java.util.EnumMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
-import org.infinispan.AbstractDelegatingAdvancedCache;
 import org.infinispan.AdvancedCache;
 import org.infinispan.Cache;
 import org.infinispan.config.Configuration;
-import org.infinispan.context.Flag;
 import org.infinispan.manager.AbstractDelegatingEmbeddedCacheManager;
 import org.infinispan.manager.CacheContainer;
 import org.infinispan.manager.EmbeddedCacheManager;
+import org.infinispan.notifications.Listener;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryActivated;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryCreated;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryInvalidated;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryLoaded;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryModified;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryPassivated;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryRemoved;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryVisited;
+import org.infinispan.notifications.cachelistener.event.Event;
 
 /**
  * @author Paul Ferraro
@@ -201,8 +218,8 @@ public class DefaultEmbeddedCacheManager extends AbstractDelegatingEmbeddedCache
         }
 
         @Override
-        public AdvancedCache<K, V> with(ClassLoader classLoader) {
-            return new ClassLoaderAwareCache<K, V>(this, classLoader);
+        public void addListener(Object listener) {
+            super.addListener(new ClassLoaderAwareListener(listener, this));
         }
 
         @Override
@@ -214,5 +231,110 @@ public class DefaultEmbeddedCacheManager extends AbstractDelegatingEmbeddedCache
         public int hashCode() {
             return this.cache.hashCode();
         }
+    }
+
+    static final Map<Event.Type, Class<? extends Annotation>> events = new EnumMap<Event.Type, Class<? extends Annotation>>(Event.Type.class);
+
+    static {
+        events.put(Event.Type.CACHE_ENTRY_ACTIVATED, CacheEntryActivated.class);
+        events.put(Event.Type.CACHE_ENTRY_CREATED, CacheEntryCreated.class);
+        events.put(Event.Type.CACHE_ENTRY_INVALIDATED, CacheEntryInvalidated.class);
+        events.put(Event.Type.CACHE_ENTRY_LOADED, CacheEntryLoaded.class);
+        events.put(Event.Type.CACHE_ENTRY_MODIFIED, CacheEntryModified.class);
+        events.put(Event.Type.CACHE_ENTRY_PASSIVATED, CacheEntryPassivated.class);
+        events.put(Event.Type.CACHE_ENTRY_REMOVED, CacheEntryRemoved.class);
+        events.put(Event.Type.CACHE_ENTRY_VISITED, CacheEntryVisited.class);
+    }
+
+    @Listener
+    public static class ClassLoaderAwareListener {
+        private final Object listener;
+        private final Map<Event.Type, List<Method>> methods = new EnumMap<Event.Type, List<Method>>(Event.Type.class);
+        private final AdvancedCache<?, ?> cache;
+
+        public ClassLoaderAwareListener(Object listener, AdvancedCache<?, ?> cache) {
+            this.listener = listener;
+            this.cache = cache;
+            for (Method method : listener.getClass().getMethods()) {
+                for (Map.Entry<Event.Type, Class<? extends Annotation>> entry : events.entrySet()) {
+                    Class<? extends Annotation> annotation = entry.getValue();
+                    if (method.isAnnotationPresent(annotation)) {
+                        List<Method> methods = this.methods.get(entry.getValue());
+                        if (methods == null) {
+                            methods = new LinkedList<Method>();
+                            this.methods.put(entry.getKey(), methods);
+                        }
+                        methods.add(method);
+                    }
+                }
+            }
+        }
+
+        @CacheEntryActivated
+        @CacheEntryCreated
+        @CacheEntryInvalidated
+        @CacheEntryLoaded
+        @CacheEntryModified
+        @CacheEntryPassivated
+        @CacheEntryRemoved
+        @CacheEntryVisited
+        public <K, V> void event(Event<K, V> event) throws Throwable {
+            List<Method> methods = this.methods.get(event.getType());
+            if (methods != null) {
+                ClassLoader cacheLoader = this.cache.getClassLoader();
+                ClassLoader contextLoader = getContextClassLoader();
+                if (cacheLoader != null) {
+                    setContextClassLoader(cacheLoader);
+                }
+                try {
+                    for (Method method : this.methods.get(event.getType())) {
+                        try {
+                            method.invoke(this.listener, event);
+                        } catch (InvocationTargetException e) {
+                            throw e.getCause();
+                        }
+                    }
+                } finally {
+                    if (cacheLoader != null) {
+                        setContextClassLoader(contextLoader);
+                    }
+                }
+            }
+        }
+
+        public int hashCode() {
+            return this.listener.hashCode();
+        }
+
+        public boolean equals(Object object) {
+            if (object == null)
+                return false;
+            if (object instanceof ClassLoaderAwareListener) {
+                ClassLoaderAwareListener listener = (ClassLoaderAwareListener) object;
+                return this.listener.equals(listener.listener);
+            }
+            return this.listener.equals(object);
+        }
+    }
+
+    static ClassLoader getContextClassLoader() {
+        PrivilegedAction<ClassLoader> action = new PrivilegedAction<ClassLoader>() {
+            @Override
+            public ClassLoader run() {
+                return Thread.currentThread().getContextClassLoader();
+            }
+        };
+        return AccessController.doPrivileged(action);
+    }
+
+    static void setContextClassLoader(final ClassLoader loader) {
+        PrivilegedAction<Void> action = new PrivilegedAction<Void>() {
+            @Override
+            public Void run() {
+                Thread.currentThread().setContextClassLoader(loader);
+                return null;
+            }
+        };
+        AccessController.doPrivileged(action);
     }
 }
