+++ clustering/infinispan/src/test/java/org/jboss/as/clustering/infinispan/DefaultEmbeddedCacheManagerTest.java
@@ -113,18 +113,6 @@ public class DefaultEmbeddedCacheManagerTest {
         assertNotSame(defaultCache, result);
         assertEquals(result, defaultCache);
         assertSame(this.subject, result.getCacheManager());
-        
-        Cache<Object, Object> cache = result;
-        AdvancedCache<Object, Object> advancedCache = cache.getAdvancedCache();
-        assertSame(cache, advancedCache);
-        AdvancedCache<Object, Object> classLoaderCache = advancedCache.with(Thread.currentThread().getContextClassLoader());
-        verify(defaultCache).addInterceptor(isA(ClassLoaderAwareCache.ClassLoaderAwareCommandInterceptor.class), eq(0));
-        assertNotSame(advancedCache, classLoaderCache);
-        
-        // Subsequent calls to with(...) cause addInterceptor() to fail
-        doThrow(new ConfigurationException("")).when(defaultCache).addInterceptor(isA(ClassLoaderAwareCache.ClassLoaderAwareCommandInterceptor.class), eq(0));
-        AdvancedCache<Object, Object> classLoaderCache2 = advancedCache.with(Thread.currentThread().getContextClassLoader());
-        assertNotSame(advancedCache, classLoaderCache2);
     }
 
     @Test
